Started networking, completed ReceiveThread
I made ReceiveThread very simple so that we can use the same class for the Server and Client. I will do the same with SendThread. This way, the Send and Receive threads will be very dumb and the controllers will be in charge of what data gets sent and what happens with the received data.